增加前端文件； 登录模块： 实现登录功能； 两次MD5加密； 集成JSR303参数校验； 增加全局异常处理器；

+++ spike_service/src/main/java/com/bitcs/dao/UserDao.java
@@ -18,7 +18,7 @@ public interface UserDao {
      * @return
      */
     @Select("select * from user where id = #{id}")
-    User getById(@Param("id") int id);
+    User getById(@Param("id") long id);
 
     @Insert("insert into user(id,name) values(#{id},#{name})")
     int insert(User user);
